Add files via upload
This application prompts the user to indicate whether they want to compare stocks in the stock market or terminate the application. If they choose to compare stocks then prompt the user to input two stocks of their choice. After, prompt the user to enter the total number of days back from the current date to perform the analysis. The results of the analysis are printed and the cycle is repeated until the user chooses to exit the application.