Add files via upload
a)	Search for content on the site.
b)    Scrape 3 pages on the site
c)    During each iteration for every row of content scraped, store the parsed and cleansed content in a new object of the class that you created in part c. Add the object to a list. 
d)  Loop through the list and use a function of  custom class to display the content of each object